backend/requirements.txt
fastapi==0.115.3
openai==1.58.1
pyarrow==17.0.0
pydantic==2.9.1
python-dotenv==1.0.1
uvicorn==0.32.0
slowapi==0.1.9
azure-core==1.32.0
azure-search-documents==11.5.2